Extract dfd types into dfd.rs
